squash changes to remove cleartext processes
