+++ public/data/food2/examples.json
@@ -4,7 +4,6 @@
     "id": "66b02aae-9c0c-48c4-aba4-e0ee191b5aca",
     "name": "Pizza bolognese Bio (350g)",
     "query": {
-      "assemblyCountry": "FR",
       "components": [
         {
           "custom": {
@@ -112,7 +111,6 @@
     "id": "7d14ebac-8ac8-4466-b334-0e9a4ba0deb5",
     "name": "Pizza bolognese FR (350g)",
     "query": {
-      "assemblyCountry": "FR",
       "components": [
         {
           "custom": {
@@ -219,7 +217,6 @@
     "id": "faff660e-2785-438e-ba1b-b04ab806b0cc",
     "name": "Pizza végétal (350g)",
     "query": {
-      "assemblyCountry": "FR",
       "components": [
         {
           "custom": {
@@ -334,7 +331,6 @@
     "id": "cb3576d3-703f-47f4-ae0c-de4f87c4c429",
     "name": "Filet de poulet FR (250g)",
     "query": {
-      "assemblyCountry": "FR",
       "components": [
         {
           "custom": {
@@ -382,7 +378,6 @@
     "id": "766efdbc-5102-4c60-bf3d-6ab32482ec28",
     "name": "Farine de blé bio FR (1kg)",
     "query": {
-      "assemblyCountry": "FR",
       "components": [
         {
           "custom": {
@@ -419,7 +414,6 @@
     "id": "230897cd-a061-47e9-a945-5acec568de8e",
     "name": "Farine de blé FR (1kg)",
     "query": {
-      "assemblyCountry": "FR",
       "components": [
         {
           "custom": {
