+++ src/Data/Textile/Stage.elm
@@ -25,6 +25,7 @@ module Data.Textile.Stage exposing
 import Area exposing (Area)
 import Data.Complement as Complement
 import Data.Country as Country exposing (Country)
+import Data.Country.Code as CountryCode
 import Data.Db exposing (Db)
 import Data.Impact as Impact exposing (Impacts)
 import Data.Process as Process exposing (Process)
@@ -494,7 +495,7 @@ airTransportDisabled : Stage -> Bool
 airTransportDisabled { country, enabled, label } =
     not enabled
         || -- Note: disallow air transport from France to France at the Making stage
-           (label == Label.Making && country.code == Country.codeFromString "FR")
+           (label == Label.Making && country.code == CountryCode.france)
 
 
 airTransportRatioToString : Split -> String
