+++ src/Page/Explore/Countries.elm
@@ -1,6 +1,7 @@
 module Page.Explore.Countries exposing (table)
 
 import Data.Country as Country exposing (Country)
+import Data.Country.Code as CountryCode
 import Data.Dataset as Dataset
 import Data.Gitbook as Gitbook
 import Data.Process as Process
@@ -22,7 +23,7 @@ import Views.Link as Link
 table : Transport.Distances -> List Country.Country -> { detailed : Bool, scope : Scope } -> Table Country String msg
 table distances countries { detailed, scope } =
     { filename = "countries"
-    , toId = .code >> Country.codeToString
+    , toId = .code >> CountryCode.toString
     , toRoute = .code >> Just >> Dataset.Countries >> Route.Explore scope
     , toSearchableWords = Country.toSearchableString >> Text.toWords
     , facets =
@@ -40,10 +41,10 @@ table distances countries { detailed, scope } =
         List.filterMap identity
             [ Just
                 { label = "Code"
-                , toValue = Table.StringValue <| .code >> Country.codeToString
+                , toValue = Table.StringValue <| .code >> CountryCode.toString
                 , toCell =
                     \country ->
-                        code [] [ text (Country.codeToString country.code) ]
+                        code [] [ text (CountryCode.toString country.code) ]
                 }
             , Just
                 { label = "Nom"
@@ -114,13 +115,13 @@ displayDistances countries distances country =
             text ""
 
 
-distancesToRows : List Country.Country -> Country.Code -> Transport.Transport -> List (Html msg) -> List (Html msg)
+distancesToRows : List Country.Country -> CountryCode.Code -> Transport.Transport -> List (Html msg) -> List (Html msg)
 distancesToRows countries countryCode transport rows =
     rows
         |> (::) (transportToRow countryCode countries transport)
 
 
-transportToRow : Country.Code -> List Country.Country -> Transport.Transport -> Html msg
+transportToRow : CountryCode.Code -> List Country.Country -> Transport.Transport -> Html msg
 transportToRow countryCode countries transport =
     let
         formatDistance length =
@@ -137,7 +138,7 @@ transportToRow countryCode countries transport =
                 |> Result.withDefault ""
                 |> title
             ]
-            [ text <| Country.codeToString countryCode ]
+            [ text <| CountryCode.toString countryCode ]
         , td [] [ formatDistance transport.air ]
         , td [] [ formatDistance transport.road ]
         , td [] [ formatDistance transport.sea ]
